0.2.3: cold storage shards + bandwidth panel + cost-based soft-release subs
Zone data storage:
- rename tiles module `zone`->`cold_tiles` and `things`->`cold_things` (symmetric
  cold-store naming); DB families tiles->cold-tiles, +cold-things.
- things Vec<u32>->Vec<u64>, adding variant:5 (32 sprites/kind) + reserved:27
  (kind:16|x:4|y:4|data:5|layer:3|variant:5|reserved:27) for stable per-thing variant.
- zone_reference u64 (server_id:16|reserved:16|zone_id:32) keys zone aggregates so a
  zone shard row can be a pipeline actor/target; edge seeds rows under it.
- worldgen derives a variant per thing; +flora thing (data/visual/biome, world/flora).

Client netcode + observability:
- debug HUD bandwidth estimate: smoothed down/up bytes/sec (main tab) + rate/total on
  the subs tab, over the existing cumulative row/reply byte tallies.
- cost-based SOFT RELEASE for zone subs (zones.rs): warm/cold/uncovered are no longer
  dropped by distance -- held until retention (held_bytes) reaches the re-fetch estimate
  (load_bytes, the LOAD_SETTLE_MS baseline burst), backstopped by the 512 capacity-LRU.
  note_update now takes the row byte length. Replaces the old update-count warmth.

Builds: edge check, cold_tiles/cold_things wasm, core check + 11 zone tests, wasm bundle,
pixijs. NOT yet deployed -- the cold-* schema change (Vec<u32>->Vec<u64>) + module
renames need `redeploy --reset`.

+++ client/core/src/api.rs
@@ -132,18 +132,17 @@ pub enum Event {
         location: u8,
         removed: bool,
     },
-    /// A subscribed zone's static terrain baseline arrived: its `ZONE_TILES` packed tile
-    /// slots (`def_id:12 | reserved:4` each, row-major). The host expands them to
-    /// renderable cells — running the DSL per `def_id` for the texture + tint — and paints
-    /// the zone's ground. Fired on every terrain row for the zone (the initial seed and any
-    /// later rewrite). Paired with [`Event::ZoneThings`] on the same terrain delivery.
-    ZoneTiles { zone_id: u32, tiles: Vec<u16> },
-    /// A subscribed zone's scattered-things baseline: the packed things in the terrain row
-    /// (`x:4 | y:4 | rotation:2 | object_id:12` each). The host expands them to sprites —
-    /// one per thing, positioned by the codec and tinted by the thing def's `:visual` (the
-    /// worldgen-scattered flora). Fired alongside [`Event::ZoneTiles`]; an empty vec means
-    /// the host clears the zone's things.
-    ZoneThings { zone_id: u32, things: Vec<u32> },
+    /// A subscribed zone's dense tile grid arrived (the `cold_tiles` module): 256 `u8`
+    /// tile-kinds, row-major (`0` = empty). The host expands them to renderable ground
+    /// cells — running the DSL per tile-kind for the texture + tint. Fired on every tiles
+    /// row for the zone (the initial seed and any later rewrite).
+    ZoneTiles { zone_id: u32, tiles: Vec<u8> },
+    /// A subscribed zone's sparse thing list arrived (the `cold_things` module): packed
+    /// things (each a `u64`: `kind:16 | x:4 | y:4 | data:5 | layer:3 | variant:5 |
+    /// reserved:27`). The host draws one sprite per thing, positioned by the codec, tinted
+    /// by the thing kind's `:visual`, and picking the kind's `variant` sprite. Fired on
+    /// every things row; an empty vec means the host clears the zone's things.
+    ZoneThings { zone_id: u32, things: Vec<u64> },
     /// A zone's subscription closed (the anchor moved it out of range, or it was
     /// evicted). The host drops that zone's entities.
     ZoneClosed { zone_id: u32 },
